requirements.txt
fastapi
uvicorn
mysql-connector-python
python-jose
python-dotenv
websockets